First end-to-end runthrough of "quick_temp_score" notebook
This changeset provides the unit-tracking ability of `Pint` for temperature scores.  Technically, we are tracking delta_degC not degC.  And with that realization, there's probably some parts of the changes that can be made simpler (because there are many math operations one cannot do with degC that one can do with delta_degC).  We leave that as an exercise for the reader.

It's also quite likely that some legit math errors have crept in that need to be chased out--nowhere does the documentation tell us what the units are in the data.  But sorting that should be the easy part.

This branch is completely orthogonal to the rmi data branch.  I did need to borrow one bugfix from there, but that's a separate world (especially considering the data units there are almost certainly different than the data units in this branch).