Temperature scores working...PR is ready for review
I sorted the problems with temperature scores and modified both input data and unit tests to work as well as can be (given outstanding issue of the test suite and pint_pandas not being too happy about the current idioms we are using).

There's much to think about in this PR.  Happy to answer questions / write up docs if/when we're ready to move forward.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>