WIP Data Vault notebook check-in
The three users are now kept separate in three separate notebooks.  There are some ugly hacks to provide the necessary separation of data via the object model.  No ugly hacks are needed for the SQL layer because that's the native language of the Data Vault.  But this checkin gives us demo-ability while we clean up the object model.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>