Integrate recently merged projections code into unit-aware code
This merge was not supposed to be that hard (one feature per branch they say), but there were many changes that overlapped and a few underlying bugs that needed to be squashed.

At this point, things that failed before (due to bad interactions with unittest) still give problems, but all the unittests  plus the new projection tests basically work as well as they did before.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>