Expand test set data and address some conditions with NULL data
The added sample data has some NULLs in the input that the tool should handle.  Some fixes address these, but not all yet.

Major fixups include preserving dtypes for columns as we winsorize data, and setting numeric_only to False in the quantile function so that we can deal with Pint quantities without triggering an error message for NA values.  (Both in base_providers.py)

Added some try wrappers in template.py to make it easier to debug when errors are thrown.

Correct some uncaught errors in how null values are processed in interfaces.py

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>