Update target_utils.py
The exponential decay of CAGR is going to operate proportional to the orders of magnitude it has to close.  If we set last to be too small, we force CAGR to spend a lot of its time chasing the asymptote.  Choosing a value of first/201 means we stop when we get within less than 1/2 of a percent, which rounds down.  That also works whether first is 1000000000 t CO2 or 1.0 Gt COt.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>