Sort production_metric and emissions_metric so they work in template test case and do not break JSON-based test cases.
Make production_metric Optional and add emission_metric (which is used in the new ITR data template).  This un-breaks older test cases based on JSON data that does not have production_metric specified.  It also relies on code to use sector (and ultimately region) to set production metric and emission_metric if not otherwise specified.

There is still much work to do to make the test case work, but the infrastructure should now be there to do so.