Enable netzero_year functionality
There was a bug in how EITargetProjector::project_ei_targets was projecting target data to 2050 absent specific targets with that as an explicit target year.  These changes fix that bug, as well as enabling the functionality of using the netzero_year field of the input template.

Also update template to use netzero_year instead of netzero_date.