test_projection WIP
These changes cut out a lot of noise about comparisons not that meaningful: previously we required the input data to supply things like ghg_s1s2 emissions at the base year, but now we collect historic data, and the tool organizes that data to fill in the singular ghg_s1s2 datapoint which is helpful to have.

There is one obvious problem, which is that input data of separate S1 and S2 emissions is not getting combined into S1S2 trajectories.  The next checkin should deal with that.  This checkin clears the clutter to make that problem easier to work on.