TemplateProviderCompany: In error log, print company id instead of name
If during the call to `_company_df_to_model` input Excel is missing
company name, error log message says:

~~~~
ERROR - (One of) the input(s) of company <NA> is invalid
~~~~

This message could be less helpful, when looking for an invalid
row in the Excel.

This commit prints company id instead, to make a problematic
row easy to find, when company name is missing:

~~~~
ERROR - (One of) the input(s) of company with ID US00130H1059 is invalid
~~~~

Company ids are used on erlier steps of Excel validation.
Also, earlier steps validate, that company ids exist for all rows.

Signed-off-by: Kirill Marinushkin <k.marinushkin@gmail.com>
Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>