Initial Scope 3 support aligned with OECM 2.0
These changes introduce breaking changes for the test suite, in particular, Scope 3 emissions are now projected according to targets and trajectories, but then merged in with S1S2 emissions, wrecking calculations that expect S3 emissions to stay in their lanes.  The reason the S3 emissions must migrate into S1S2 is related to how the "production-centric" benchmarks are constructed (they have scope 3 = 0 for all but cement).  Since methodology is foremost, and since the benchmarks speak to how to handle S3, we are following that until we have benchmarks that keep S3 separate and nonzero.

Signed-off-by: MichaelTiemannOSC  <mtiemann@os-climate.org>
Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>