quick_template_score_calc.ipynb: Removed unused github token
GitHub token was required for this notebook, but never used.
At the same time, not all ITR users have a GitHub token,
which made this notebook not executable for them.

When i removed it, it's absense didn't effect execution
of the notebook. Which means, that it can be removed safely,
and make the notebook more accessible for ITR users.

Signed-off-by: Kirill Marinushkin <k.marinushkin@gmail.com>
Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>