Implement exceedance years; Reduce warnings
The tool now adds "exceedance year" analysis to the enhanced portfolio.  The exceedance year is when whatever the budget was according to the benchmark, the company has surpassed the budget and has no future target that can return them to alignment.

Also greatly reduced the number of warnings the tool generates, due to pytest, pandas, and pint.

Also fixed several idiomatically bad practices such as using df.apply when df[col].map would do as well or better.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>