Exceedance Year enhancements for ITR_UI
The exceedance year functionality needs some feedback from users and we need to decide whether to make it part of the tool or a demo of the tool (in an example notebook, for example).  But the implementation of the calculation is improved in this version (simpler and better than initial commit).

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>