Allocate emissions across multiple sectors
When a company reports production figures across multiple sectors, total emissions should somehow be allocated to those production activities so they can be analyzed and scored.

These changes allow companies to report muliple types of production and the tool to create synthetic companies that have benchmark-aligned and allocated emissions attributed.  But they do not yet drive that logic through to targets.  That will come in a future commit.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>