Update ITR_UI.py
Add new Dash Slider to control TARGET_YEAR.

Note this also carries with it a flip-flop on the convention of whether we should pass scopes=None or scopes=[some particular set of scopes] as the default in the source tree.  It will continue to flip and flop until we decide what's best.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>