Display benchmark budget and temp score
It's no surprise that a 1.5˚C benchmark would have a 1.5˚C implied temperature score at the netzero point of the benchmark.  But when selecting target years less than the target year, it's helpful to know "what's the temperature score of the benchmark at the new target year"?  And "what is the total budget used between the BASE_YEAR and the TARGET_YEAR.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>