update_chart: Input->State to avoid race condition
In Dash, the first time through, all inputs must be available before a function can run.  Thereafter, *ANY* subsequent input can trigger a function to run.  Changing a the Scope parameter was causing both a recalculation of the portfolio dataframe as well as triggering update_graph, which also depends on the portfolio.  If update_graph was triggered by Scope before the new portfolio was ready, it would fail.  By changing the Input to State, only a fresh portfolio (that was already aligned with the state of Scope) would trigger update_graph.

Also, remove an errant breakpoint that was left behind previously.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>