Don't introduce global variables to Dash
The default ProjectionControls used for ProviderCompany data is the class ProjectionControls (not an instance of that class).  As such, their fields are effectively global variables, which is a No-No for Dash.  So we request a fresh instance at the start, and this can be updated and used by Dash.

Also, by treating "target_year" as a state variable for "warehouse-eibm" it makes "target-year" persistent across different benchmark selections...a nice feature.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>