Fix all-NA rows of data in V1 template
When translating Template V1 data, it could happen that all ghg_s3 data is NA, which would create an all-NA row of emissions data.  This was properly handled in Template V2 processing, but not in Template V1 processing.  Changing this required fixing the test case `test_temp_score_from_excel_data` in test_template_provider.py (which allowed missing data to be represented by the default temperature score 3.2˚C).

Once fixed, it revealed a second problem, which was handling mixed float64 and UFloat data across target and trajectory projections.

And finally, a third problem, which was that column order for the output spreadsheet was not stable (due to using a set intersection operation, which hashes values and potentially changes their order).

Now, all fixed.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>