Fix latent errors when inferring S2 emissions
In the course of inferring Scope 2 emissions, it was discovered that _validate_projected_trajectories was not removing companies with invalid trajectories from further analysis.  It was also discovered that callers expected _validate_projected_trajectories to do this work, whereas the function itself was trying to return the validated list for the caller to then use.  Now the function is defined and used consistently.

project_ei_trajectories would fail with e a KeyError due to a typo in the key definition.

New math introduced in a previous commit would cause _get_cumulative_emissions to create NaNs with nan uncertainties.  Elsewhere we've told ourselves that nans with uncertainties will always be of the form `nan +/- 0.0` so we need to normalize these nans to what we expect (or be looser in terms of what sorts of nans we allow).

In _convert_from_template_company_data, remove an extraneous metric from logger error messages to make output more compact.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>