Fix target unit alignment error
It can happen that disclosure units and EI benchmark units are not perfectly aligned.  Pint can do some automatic conversions, but it is not aggressive in converting units for normal algebraic operations.  By adding value.to(other_units) conversions when needed, we bridge those gaps.

In the course of that one-line change in target projections, a second unit conversion issue was revealed and fixed, along with some other very small cleanups.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>