Create proper MI in _get_intensity_benchmarks
Also the Pandas release candidate enforces that we cannot change the `names` property of a DataFrame's columns, so we regenerate the column's multiindex with the names we want in `_get_intensity_benchmarks`

Also add a stop-gap testing fix.  The latest changes to `develop` have a better fix we can use.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>