Fix/document umean calculation
The umean calculation is primarily meant to take a bunch of nominal values and compute the mean and standard error as the uncertainty.  But it also implements the concept of "the uncertainty of the mean", which is very different.  We "compromise" by implementing the former if any incoming values have no or zero uncertainty, and the latter if all values arrive with non-zero error terms.

Also, remove an errant breakpoint in _get_cumulative_emissions, and make the function report better error messages if inputs can't be made into PintArray-backed Series.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>