Ignore OSV:PYSEC-2023-163
This vulnerability invalidly flags numexpr for a problem that is actually due to LangChain.  Nothing in ITR accesses `numexpr` in a way that would trigger this vulnerability.  Of course a naive or malicious user could create a Jupyter Notebook that would open its server to 3rd-party attack, but this is true of any Python environment (that already provides `eval`).

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>