Chore: Push datetime into GITHUB_ENV
Signed-off-by: ModeSevenIndustrialSolutions <93649628+ModeSevenIndustrialSolutions@users.noreply.github.com>

This specific commit exists only to shut up the git rebase process that issued confusing guidance
about what to do about a file that exists in the history but doesn't exist any longer.  Somehow
deleting the non-existent file greatly disturbed the dead ancentors of this file, which will
haunt at least the next 58 commits we have to manually merge otherwise.