First attempt to merge pydantic changes post lint
The linting process that cleaned up the ITR python code in general was helpful, but it interacted very badly with a large pending commit to support pydantic 2.0.  As a result, I had to manually vet all the lint-related changes and to do my best to pick and choose whether to edit the linted code or essentially re-lint the new pydantic code.

But also this is my first attempt to use the new world of tox, etc., and I cannot myself run local tests yet.  I'm therefore creating this test as a checkpoint which we can use to sort how how to make things work in a new directory structure (src/) and a new way to activating and using virtual environments (tox) with which I am completely unfamiliar, and which are yet to be documented in the README as a series of steps that I can follow from the command line.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>