Reconcile changes in forklift update (tests pass)
The forklift update was a best-effort to re-establish a working environment after some forking problems, and problems that inhibited local testing.  Now that we have that all sorted, these changes should bring everything back into line.

We still have the nasty problem that pre-commit is not part of our standard environment, and that making it part of the environment brings in nearly 2000 fresh files from the command `npm install eslint-plugin-toml@latest --save-dev`

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>