Update pyproject.toml
Update osc-ingest-tools so that we can use SqlAlchemy 2.0.

Other OS-Climate repos can decide when and if they want to follow suit.  Since these tools are anticipated to run in their own virtual environments (at least for now), there shouldn't be a lot of version conflict.  Data pipelines that do feed ITR (and might be running in the same virtual environment) can be trivially updated by replacing patterns of the form:

```
    engine.execute(\"select * from essd.regions\").fetchall()
```
with
```
    with engine.begin() as cxn:\n",
        qres = cxn.execute(text(\"select * from essd.regions\")).fetchall()
```

Note the call to `text` as we cannot pass raw strings to `execute` (to better protect against SQL injection attacks).

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>