Update template.py
The existing code didn't function correctly when all production values are NA for the `BASE_YEAR`.  We have a decision to make whether we want to backfill production values from later years to the base year, but in the meantime, at least we now handle the case better, and can protect the tool from sending empty dataframes to the UI tool.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>