Rewrite list-map-zip for mypy
Guido, and hence Python, do not favor lamba expressions.  Among other things, lambdas can confuse mypy and there's no great syntactic sugar to fix this.

We fix this two ways.  In the simple case, we can provide a type signature using typing.cast.  The a more complex case, we rewrite a nested comprehension that was written to create three lists as a simple loop to create those three lists with proper type signatures.  And in the process, simplify some old code that is no longer relevant.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>