Mostly complete WIP commit
Tons of changes to bring Data Vault into a state of compliance with the rest of the project.  There's still much to do, but we pass both test suites and ITR_DV.py mostly works.

These changes also depend on a change to osc_units.py, but I'm putting that in a parallel check-in that is independent of these changes.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>