WIP changes automobile equivalences (veh=150k kpm)
Benchmark units such as passenger km are all well and good, but when reports are made in vehicles produced, it's a pain for users to convert from vehicle counts to passenger kilometers.  This allows the specification of production equivalence relationships (first in the case of auto sector, then generalizable to other sectors).

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>